+++ migrations/0008_create_amenities.up.sql
@@ -3,8 +3,8 @@ CREATE TABLE amenities(
   created_at timestamp without time zone not null,
   type varchar not null,
   name varchar check (type <> 'custom' or name is not null),
-  property_id integer references properties,
-  unit_id integer references units check (property_id is not null or unit_id is not null)
+  property_id bigint references properties,
+  unit_id bigint references units check (property_id is not null or unit_id is not null)
 );
 
 CREATE UNIQUE INDEX amenities_name_uniq_by_property ON amenities(property_id, name);
