+++ migrations/0009_create_bookings.up.sql
@@ -1,15 +1,15 @@
 CREATE TABLE bookings(
   id bigserial primary key not null,
   created_at timestamp without time zone not null,
-  user_id integer not null references users,
-  unit_id integer not null references units,
+  user_id bigint not null references users,
+  unit_id bigint not null references units,
   check_in date not null,
   check_out date not null,
   message text,
-  night_price_cents integer not null,
-  nights integer not null check (nights > 0),
-  service_fee_cents integer not null,
-  total_cents integer not null
+  night_price_cents int not null,
+  nights int not null check (nights > 0),
+  service_fee_cents int not null,
+  total_cents int not null
 );
 
 CREATE INDEX bookings_user_id ON bookings(user_id);
